+++ src/routes/auth.routes.test.ts
@@ -6,8 +6,8 @@
  */
 import request from "supertest";
 import mongoose from "mongoose";
-import app from "../src/server";
-import User from "../src/models/user.model";
+import app from "../server";
+import User from "../models/user.model";
 
 describe("Auth Routes Integration Tests", () => {
   beforeAll(async () => {